+++ src/de/bassadin/mathsimsemgrp1_1/Objects2D/Quad.java
@@ -6,7 +6,7 @@ public class Quad extends Object2D {
     protected int width;
     protected int height;
 
-    public Quad(int posX, int posY, int width, int height) {
+    public Quad(double posX, double posY, int width, int height) {
         super(posX, posY);
         this.width = width;
         this.height = height;
@@ -14,7 +14,7 @@ public class Quad extends Object2D {
 
     public void draw(Graphics g) {
         g.setColor(color);
-        g.fillRect(posX, posY, width, height);
+        g.fillRect((int)posX, (int)posY, width, height);
     }
 
     public int getWidth() {
